Separate CameraFrame into its own file
